+++ packages/design-system-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.4.2-pr-24.2 (2022-10-05)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.4.2-pr-24.1 (2022-10-05)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.4.2-pr-24.0 (2022-10-05)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
 ## 1.4.1 (2022-09-30)
 
 **Note:** Version bump only for package @geovistory/design-system-react